requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==2.2.5
numpy==1.15.4
pip==9.0.3
pyparsing==2.4.7
pyserial==3.0
python-dateutil==2.8.1
pytz==2020.1
setuptools==12.0.5
six==1.15.0
pyqt5==5.5.1
